+++ src/Entity/ApiGenerator/NowPlayingApiGenerator.php
@@ -164,7 +164,7 @@ final class NowPlayingApiGenerator
         $np->station = $this->stationApiGenerator->__invoke($station, $baseUri);
         $np->listeners = new Listeners();
 
-        $songObj = Song::createOffline();
+        $songObj = Song::createOffline($station->getBrandingConfig()->getOfflineText());
 
         $offlineApiNowPlaying = new CurrentSong();
         $offlineApiNowPlaying->sh_id = 0;