+++ src/Entity/Song.php
@@ -12,6 +12,8 @@ class Song implements SongInterface
 {
     use Traits\HasSongFields;
 
+    public final const OFFLINE_SONG_ID = '5a6a865199cf5df73b1417326d2ff24f';
+
     public function __construct(?SongInterface $song = null)
     {
         if (null !== $song) {
@@ -62,6 +64,10 @@ class Song implements SongInterface
 
         $songText = str_replace($removeChars, '', $songText);
 
+        if (empty($songText)) {
+            return self::OFFLINE_SONG_ID;
+        }
+
         $hashBase = mb_strtolower($songText, 'UTF-8');
         return md5($hashBase);
     }
@@ -104,8 +110,10 @@ class Song implements SongInterface
         return self::createFromNowPlayingSong($currentSong);
     }
 
-    public static function createOffline(): self
+    public static function createOffline(?string $text = null): self
     {
-        return self::createFromText('Station Offline');
+        $song = self::createFromText($text ?? 'Station Offline');
+        $song->setSongId(self::OFFLINE_SONG_ID);
+        return $song;
     }
 }